Refactor from_ugrid_conventions to model
