+++ src/parcels/_core/model.py
@@ -15,8 +15,14 @@ from parcels._core.uxgrid import UxGrid
 from parcels._core.xgrid import _DEFAULT_XGCM_KWARGS, XGrid
 from parcels._logger import logger
 from parcels._typing import Mesh
+from parcels.convert import _ds_rename_using_standard_names
 from parcels.interpolators import (
     CGrid_Velocity,
+    Ux_Velocity,
+    UxConstantFaceConstantZC,
+    UxConstantFaceLinearZF,
+    UxLinearNodeConstantZC,
+    UxLinearNodeLinearZF,
     XLinear,
     XLinear_Velocity,
 )
@@ -118,7 +124,37 @@ class UnstructuredModel(Model):
         self.data = data
         self.grid = grid
 
-    def construct_fields(self) -> list[Field | VectorField]: ...
+    def construct_fields(self) -> list[Field | VectorField]:
+        ds = self.data
+        grid = self.grid
+        fields = {}
+        if "U" in ds.data_vars and "V" in ds.data_vars:
+            fields["U"] = Field("U", ds["U"], grid, _select_uxinterpolator(ds["U"]))
+            fields["V"] = Field("V", ds["V"], grid, _select_uxinterpolator(ds["V"]))
+            fields["UV"] = VectorField("UV", fields["U"], fields["V"], vector_interp_method=Ux_Velocity)
+
+            if "W" in ds.data_vars:
+                fields["W"] = Field("W", ds["W"], grid, _select_uxinterpolator(ds["W"]))
+                fields["UVW"] = VectorField(
+                    "UVW", fields["U"], fields["V"], fields["W"], vector_interp_method=Ux_Velocity
+                )
+
+        for varname in set(ds.data_vars) - set(fields.keys()):
+            fields[varname] = Field(str(varname), ds[varname], grid, _select_uxinterpolator(ds[varname]))
+
+        return list(fields.values())
+
+    @classmethod
+    def from_ugrid_conventions(cls, ds: ux.UxDataset, mesh: str = "spherical"):
+        ds_dims = list(ds.dims)
+        if not all(dim in ds_dims for dim in ["time", "zf", "zc"]):
+            raise ValueError(
+                f"Dataset missing one of the required dimensions 'time', 'zf', or 'zc' for uxDataset. Found dimensions {ds_dims}"
+            )
+
+        grid = UxGrid(ds.uxgrid, z=ds.coords["zf"], mesh=mesh)
+        ds = _discover_ux_U_and_V(ds)
+        return cls(ds, grid)
 
 
 # TODO: Refactor later into something like `parcels._metadata.discover(dataset)` helper that can be used to discover important metadata like this. I think this whole metadata handling should be refactored into its own module.
@@ -148,6 +184,85 @@ def _is_coordinate_in_degrees(da: xr.DataArray) -> bool:
     if isinstance(units, str) and "degree" in units.lower():
         return True
     return False
+
+
+def _discover_ux_U_and_V(ds: ux.UxDataset) -> ux.UxDataset:
+    # Common variable names for U and V found in UxDatasets
+    common_ux_UV = [("unod", "vnod"), ("u", "v")]
+    common_ux_W = ["w"]
+
+    if "W" not in ds:
+        for common_W in common_ux_W:
+            if common_W in ds:
+                ds = _ds_rename_using_standard_names(ds, {common_W: "W"})
+                break
+
+    if "U" in ds and "V" in ds:
+        return ds  # U and V already present
+    elif "U" in ds or "V" in ds:
+        raise ValueError(
+            "Dataset has only one of the two variables 'U' and 'V'. Please rename the appropriate variable in your dataset to have both 'U' and 'V' for Parcels simulation."
+        )
+
+    for common_U, common_V in common_ux_UV:
+        if common_U in ds:
+            if common_V not in ds:
+                raise ValueError(
+                    f"Dataset has variable with standard name {common_U!r}, "
+                    f"but not the matching variable with standard name {common_V!r}. "
+                    "Please rename the appropriate variables in your dataset to have both 'U' and 'V' for Parcels simulation."
+                )
+            else:
+                ds = _ds_rename_using_standard_names(ds, {common_U: "U", common_V: "V"})
+                break
+
+        else:
+            if common_V in ds:
+                raise ValueError(
+                    f"Dataset has variable with standard name {common_V!r}, "
+                    f"but not the matching variable with standard name {common_U!r}. "
+                    "Please rename the appropriate variables in your dataset to have both 'U' and 'V' for Parcels simulation."
+                )
+            continue
+
+    return ds
+
+
+def _select_uxinterpolator(da: ux.UxDataArray):
+    """Selects the appropriate uxarray interpolator for a given uxarray UxDataArray"""
+    supported_uxinterp_mapping = {
+        # (zc,n_face): face-center laterally, layer centers vertically — piecewise constant
+        "zc,n_face": UxConstantFaceConstantZC,
+        # (zc,n_node): node/corner laterally, layer centers vertically — barycentric lateral & piecewise constant vertical
+        "zc,n_node": UxLinearNodeConstantZC,
+        # (zf,n_node): node/corner laterally, layer interfaces vertically — barycentric lateral & linear vertical
+        "zf,n_node": UxLinearNodeLinearZF,
+        # (zf,n_face): face-center laterally, layer interfaces vertically — piecewise constant lateral & linear vertical
+        "zf,n_face": UxConstantFaceLinearZF,
+    }
+    # Extract only spatial dimensions, neglecting time
+    da_spatial_dims = tuple(d for d in da.dims if d not in ("time",))
+    if len(da_spatial_dims) != 2:
+        raise ValueError(
+            "Fields on unstructured grids must have two spatial dimensions, one vertical (zf or zc) and one lateral (n_face, n_edge, or n_node)"
+        )
+
+    # Construct key (string) for mapping to interpolator
+    # Find vertical and lateral tokens
+    vdim = None
+    ldim = None
+    for d in da_spatial_dims:
+        if d in ("zf", "zc"):
+            vdim = d
+        if d in ("n_face", "n_node"):
+            ldim = d
+    # Map to supported interpolators
+    if vdim and ldim:
+        key = f"{vdim},{ldim}"
+        if key in supported_uxinterp_mapping.keys():
+            return supported_uxinterp_mapping[key]
+
+    return None
 
 
 def _is_agrid(ds: xr.Dataset) -> bool:
